util.go 内の関数を collector.go にマージする
